+++ crates/daoyi-entities-system/src/entity/system_dict_type.rs
@@ -7,17 +7,17 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "system_dict_type")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub name: String,
     pub r#type: String,
-    pub status: i16,
+    pub status: i32,
     pub remark: Option<String>,
     pub creator: Option<String>,
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub deleted_time: Option<DateTime>,
 }
