+++ crates/daoyi-entities-system/src/entity/system_notice.rs
@@ -7,18 +7,18 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "system_notice")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub title: String,
     #[sea_orm(column_type = "Text")]
     pub content: String,
-    pub r#type: i16,
-    pub status: i16,
+    pub r#type: i32,
+    pub status: i32,
     pub creator: Option<String>,
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
