+++ crates/daoyi-entities-system/src/entity/system_notify_message.rs
@@ -7,10 +7,10 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "system_notify_message")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub user_id: i64,
-    pub user_type: i16,
+    pub user_type: i32,
     pub template_id: i64,
     pub template_code: String,
     pub template_nickname: String,
@@ -23,7 +23,7 @@ pub struct Model {
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
