+++ crates/daoyi-entities-system/src/entity/system_social_user_bind.rs
@@ -7,17 +7,17 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "system_social_user_bind")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub user_id: i64,
-    pub user_type: i16,
-    pub social_type: i16,
+    pub user_type: i32,
+    pub social_type: i32,
     pub social_user_id: i64,
     pub creator: Option<String>,
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
