+++ crates/daoyi-module-demo/src/entity/yudao_demo01_contact.rs
@@ -7,10 +7,10 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "yudao_demo01_contact")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub name: String,
-    pub sex: i16,
+    pub sex: i32,
     pub birthday: DateTime,
     pub description: String,
     pub avatar: Option<String>,
@@ -18,7 +18,7 @@ pub struct Model {
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
