+++ crates/daoyi-module-demo/src/entity/yudao_demo03_student.rs
@@ -7,17 +7,17 @@ use serde::{Deserialize, Serialize};
 #[sea_orm(schema_name = "dev", table_name = "yudao_demo03_student")]
 #[serde(rename_all = "camelCase")]
 pub struct Model {
-    #[sea_orm(primary_key, auto_increment = false)]
+    #[sea_orm(primary_key, auto_increment = true)]
     pub id: i64,
     pub name: String,
-    pub sex: i16,
+    pub sex: i32,
     pub birthday: DateTime,
     pub description: String,
     pub creator: Option<String>,
     pub create_time: DateTime,
     pub updater: Option<String>,
     pub update_time: DateTime,
-    pub deleted: i16,
+    pub deleted: i32,
     pub tenant_id: i64,
 }
